clean blog pages
